Basic Network connect added.
Note: The server project is not added to the git yet. It will be added once enough work is done and the client-server code is ready for further testing with other parts of the game. Until then the network code is mainly only my concern with some tidbits for Jasdip to code the Unit class.